evaluation cleanup

+++ bash/evaluate/nc/EUCLIDEAN_NC.sh
@@ -3,7 +3,7 @@
 #SBATCH --job-name=EUCLIDNC
 #SBATCH --output=EUCLIDNC_%A_%a.out
 #SBATCH --error=EUCLIDNC_%A_%a.err
-#SBATCH --array=0-3599
+#SBATCH --array=0-2999
 #SBATCH --time=10-00:00:00
 #SBATCH --ntasks=1
 #SBATCH --mem=5G
@@ -11,7 +11,7 @@
 datasets=(cora_ml citeseer ppi pubmed mit)
 dims=(5 10 25 50)
 seeds=({0..29})
-methods=(abrw attrpure deepwalk tadw aane sagegcn)
+methods=(attrpure deepwalk tadw aane sagegcn)
 exp=nc_experiment
 
 num_datasets=${#datasets[@]}
@@ -31,7 +31,6 @@ method=${methods[$method_id]}
 
 echo $dataset $dim $seed $method
 
-
 data_dir=datasets/${dataset}
 edgelist=${data_dir}/edgelist.tsv.gz
 features=${data_dir}/feats.csv.gz 
@@ -41,15 +40,20 @@ embedding_dir=$(echo ../OpenANE/embeddings/${dataset}/nc_experiment/${dim}/${met
 test_results=$(printf "test_results/${dataset}/${exp}/${method}/dim=%03d/" ${dim})
 echo $embedding_dir
 
-args=$(echo --edgelist ${edgelist} --labels ${labels} \
-    --dist_fn euclidean \
-    --embedding ${embedding_dir} --seed ${seed} \
-    --test-results-dir ${test_results})
-echo $args
+if [ ! -f ${test_results}/${seed}.pkl ]
+then
+    args=$(echo --edgelist ${edgelist} --labels ${labels} \
+        --dist_fn euclidean \
+        --embedding ${embedding_dir} --seed ${seed} \
+        --test-results-dir ${test_results})
+    echo $args
 
-module purge
-module load bluebear
-module load Python/3.6.3-iomkl-2018a
-pip install --user numpy pandas networkx scikit-learn scikit-multilearn
+    module purge
+    module load bluebear
+    module load Python/3.6.3-iomkl-2018a
+    pip install --user numpy pandas networkx scikit-learn scikit-multilearn
 
-python evaluate_nc.py ${args}
+    python evaluate_nc.py ${args}
+else 
+    echo ${test_results}/${seed}.pkl already exists 
+fi
