+++ bash/evaluate/nc/NK_NC.sh
@@ -44,9 +44,15 @@ args=$(echo --edgelist ${edgelist} --labels ${labels} \
     --test-results-dir ${test_results})
 echo $args
 
-module purge
-module load bluebear
-module load Python/3.6.3-iomkl-2018a
-pip install --user numpy pandas networkx scikit-learn scikit-multilearn
 
-python evaluate_nc.py ${args}
+if [ ! -f ${test_results}/${seed}.pkl ]
+then
+    module purge
+    module load bluebear
+    module load Python/3.6.3-iomkl-2018a
+    pip install --user numpy pandas networkx scikit-learn scikit-multilearn
+
+    python evaluate_nc.py ${args}
+else 
+    echo ${test_results}/${seed}.pkl already exists 
+fi
